+++ tests/integration/test_allometry.py
@@ -1,10 +1,20 @@
 import os.path
 import shutil
+import sys
+
+import pytest
 
 from pharmpy.internals.fs.cwd import chdir
 from pharmpy.model import Model
 from pharmpy.tools import read_modelfit_results, run_tool
 from pharmpy.workflows import ModelDatabase
+
+tflite_condition = (
+    sys.version_info >= (3, 12)
+    and sys.platform == 'win32'
+    or sys.version_info >= (3, 12)
+    and sys.platform == 'darwin'
+)
 
 
 def test_allometry(tmp_path, testdata):
@@ -34,3 +44,18 @@ def test_allometry(tmp_path, testdata):
                 if line[:6] == '$DATA ':
                     assert line == f'$DATA ..{sep}.datasets{sep}input_model.csv IGNORE=@\n'
                     break
+
+
+@pytest.mark.skipif(tflite_condition, reason="Skipping tests requiring tflite for Python 3.12")
+def test_run_allometry(tmp_path, load_model_for_test, testdata):
+    model = load_model_for_test(testdata / 'nonmem' / 'models' / 'pheno5.mod')
+    results = read_modelfit_results(testdata / 'nonmem' / 'models' / 'pheno5.mod')
+    with chdir(tmp_path):
+        res = run_tool(
+            'allometry',
+            model=model,
+            results=results,
+            allometric_variable='WGT',
+            estimation_tool='dummy',
+        )
+        assert len(res.summary_models) == 2
