+++ tests/integration/test_amd.py
@@ -163,3 +163,27 @@ def test_amd_dollar_design(tmp_path, testdata):
         assert len(res.summary_tool) == len(subrundir)
         assert len(res.summary_models) >= 1
         assert len(res.summary_individuals_count) >= 1
+
+
+@pytest.mark.filterwarnings(
+    'ignore:.*Adjusting initial estimates to create positive semidefinite omega/sigma matrices.',
+    'ignore::UserWarning',
+    'ignore::FutureWarning',
+    'ignore::Warning',
+)
+def test_amd_dummy(tmp_path, testdata):
+    with chdir(tmp_path):
+        shutil.copy2(testdata / 'nonmem' / 'models' / 'moxo_simulated_amd.csv', '.')
+        shutil.copy2(testdata / 'nonmem' / 'models' / 'moxo_simulated_amd.datainfo', '.')
+        input = 'moxo_simulated_amd.csv'
+        run_amd(
+            input,
+            modeltype='basic_pk',
+            administration='oral',
+            search_space='PERIPHERALS(1)',
+            strategy='SIR',
+            occasion='VISI',
+            strictness='minimization_successful or rounding_errors',
+            retries_strategy='skip',
+            estimation_tool='dummy',
+        )
