+++ tests/tools/test_amd.py
@@ -397,27 +397,3 @@ def _model_count(rundir: Path):
             ((rundir / 'models').iterdir()),
         )
     )
-
-
-@pytest.mark.filterwarnings(
-    'ignore:.*Adjusting initial estimates to create positive semidefinite omega/sigma matrices.',
-    'ignore::UserWarning',
-    'ignore::FutureWarning',
-    'ignore::Warning',
-)
-def test_amd(tmp_path, testdata):
-    with chdir(tmp_path):
-        shutil.copy2(testdata / 'nonmem' / 'models' / 'moxo_simulated_amd.csv', '.')
-        shutil.copy2(testdata / 'nonmem' / 'models' / 'moxo_simulated_amd.datainfo', '.')
-        input = 'moxo_simulated_amd.csv'
-        run_amd(
-            input,
-            modeltype='basic_pk',
-            administration='oral',
-            search_space='PERIPHERALS(1)',
-            strategy='SIR',
-            occasion='VISI',
-            strictness='minimization_successful or rounding_errors',
-            retries_strategy='skip',
-            estimation_tool='dummy',
-        )
