+++ tests/tools/test_covsearch.py
@@ -1,23 +1,12 @@
-import sys
-
 import pytest
 
-from pharmpy.internals.fs.cwd import chdir
 from pharmpy.modeling import add_covariate_effect, get_covariate_effects, remove_covariate_effect
-from pharmpy.tools import read_modelfit_results, run_tool
 from pharmpy.tools.covsearch.tool import (
     create_workflow,
     filter_search_space_and_model,
     validate_input,
 )
 from pharmpy.workflows import Workflow
-
-tflite_condition = (
-    sys.version_info >= (3, 12)
-    and sys.platform == 'win32'
-    or sys.version_info >= (3, 12)
-    and sys.platform == 'darwin'
-)
 
 MINIMAL_INVALID_MFL_STRING = ''
 MINIMAL_VALID_MFL_STRING = 'LET(x, 0)'
@@ -176,19 +165,3 @@ def test_covariate_filtering(load_model_for_test, testdata):
     eff, filtered_model = filter_search_space_and_model(search_space, model)
     assert len(get_covariate_effects(filtered_model)) == 2
     assert len(eff) == 0
-
-
-@pytest.mark.skipif(tflite_condition, reason="Skipping tests requiring tflite for Python 3.12")
-def test_default_str(tmp_path, load_model_for_test, testdata):
-    model = load_model_for_test(testdata / 'nonmem' / 'models' / 'mox2.mod')
-    results = read_modelfit_results(testdata / 'nonmem' / 'models' / 'mox2.mod')
-    with chdir(tmp_path):
-        run_tool(
-            'covsearch',
-            'LET(CONTINUOUS, [AGE, WT]); LET(CATEGORICAL, SEX)\n'
-            'COVARIATE?([CL, MAT, VC], @CONTINUOUS, exp, *)\n'
-            'COVARIATE?([CL, MAT, VC], @CATEGORICAL, cat, *)',
-            results=results,
-            model=model,
-            estimation_tool='dummy',
-        )
