+++ tests/tools/test_iovsearch.py
@@ -1,11 +1,7 @@
-import sys
-
 import pytest
 
 from pharmpy.basic import Expr
-from pharmpy.internals.fs.cwd import chdir
 from pharmpy.modeling import add_iov, fix_parameters, remove_iov
-from pharmpy.tools import read_modelfit_results, run_iovsearch
 from pharmpy.tools.iovsearch.tool import (
     _get_iiv_etas_with_corresponding_iov,
     _get_nonfixed_iivs,
@@ -13,13 +9,6 @@ from pharmpy.tools.iovsearch.tool import (
     validate_input,
 )
 from pharmpy.workflows import Workflow
-
-tflite_condition = (
-    sys.version_info >= (3, 12)
-    and sys.platform == 'win32'
-    or sys.version_info >= (3, 12)
-    and sys.platform == 'darwin'
-)
 
 
 def S(x):
@@ -124,13 +113,3 @@ def test_validate_input_raises(
 
     with pytest.raises(exception, match=match):
         validate_input(**kwargs)
-
-
-@pytest.mark.skipif(tflite_condition, reason="Skipping tests requiring tflite for Python 3.12")
-def test_default_mox2(tmp_path, load_model_for_test, testdata):
-    model = load_model_for_test(testdata / 'nonmem' / 'models' / 'mox1.mod')
-    results = read_modelfit_results(testdata / 'nonmem' / 'models' / 'mox1.mod')
-    with chdir(tmp_path):
-        run_iovsearch(
-            'VISI', rank_type='ofv', results=results, model=model, estimation_tool='dummy'
-        )
